+++ pyppbox/config/myconfig.py
@@ -4,16 +4,16 @@
 #   Copyright (C) 2025 UMONS-Numediart                                      #
 #                                                                           #
 #   This program is free software: you can redistribute it and/or modify    #
-#   it under the terms of the GNU General Public License as published by    #
-#   the Free Software Foundation, either version 3 of the License, or       #
-#   (at your option) any later version.                                     #
+#   it under the terms of the GNU Affero General Public License as          #
+#   published by the Free Software Foundation, either version 3 of the      #
+#   License, or (at your option) any later version.                         #
 #                                                                           #
 #   This program is distributed in the hope that it will be useful,         #
 #   but WITHOUT ANY WARRANTY; without even the implied warranty of          #
 #   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           #
-#   GNU General Public License for more details.                            #
+#   GNU Affero General Public License for more details.                     #
 #                                                                           #
-#   You should have received a copy of the GNU General Public License       #
+#   You should have received a copy of the GNU Affero General Public License#
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.  #
 #                                                                           #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
@@ -1158,16 +1158,16 @@ class MyCFGHeaders(object):
                 "::    Copyright (C) 2025 UMONS-Numediart\n"
                 "::\n"
                 "::    This program is free software: you can redistribute it and/or modify\n"
-                "::    it under the terms of the GNU General Public License as published by\n"
+                "::    it under the terms of the GNU Affero General Public License as published by\n"
                 "::    the Free Software Foundation, either version 3 of the License, or\n"
                 "::    (at your option) any later version.\n"
                 "::\n"
                 "::    This program is distributed in the hope that it will be useful,\n"
                 "::    but WITHOUT ANY WARRANTY; without even the implied warranty of\n"
                 "::    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the\n"
-                "::    GNU General Public License for more details.\n"
+                "::    GNU Affero General Public License for more details.\n"
                 "::\n"
-                "::    You should have received a copy of the GNU General Public License\n"
+                "::    You should have received a copy of the GNU Affero General Public License\n"
                 "::    along with this program.  If not, see <https://www.gnu.org/licenses/>.\n\n\n")
         return header