+++ pyppbox/gui/ui_launcher.py
@@ -4,16 +4,16 @@
 #   Copyright (C) 2025 UMONS-Numediart                                      #
 #                                                                           #
 #   This program is free software: you can redistribute it and/or modify    #
-#   it under the terms of the GNU General Public License as published by    #
-#   the Free Software Foundation, either version 3 of the License, or       #
-#   (at your option) any later version.                                     #
+#   it under the terms of the GNU Affero General Public License as          #
+#   published by the Free Software Foundation, either version 3 of the      #
+#   License, or (at your option) any later version.                         #
 #                                                                           #
 #   This program is distributed in the hope that it will be useful,         #
 #   but WITHOUT ANY WARRANTY; without even the implied warranty of          #
 #   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           #
-#   GNU General Public License for more details.                            #
+#   GNU Affero General Public License for more details.                     #
 #                                                                           #
-#   You should have received a copy of the GNU General Public License       #
+#   You should have received a copy of the GNU Affero General Public License#
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.  #
 #                                                                           #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
@@ -242,7 +242,7 @@ class Ui_PYPPBOXLauncher(object):
 
     def retranslateUi(self, PYPPBOXLauncher):
         _translate = QtCore.QCoreApplication.translate
-        title = "PYPPBOX (GPLV3+) - VERSION " + getVersionString()
+        title = "PYPPBOX (AGPLV3+) - VERSION " + getVersionString()
         PYPPBOXLauncher.setWindowTitle(_translate("PYPPBOXLauncher", title))
         self.launch_pushButton.setText(_translate("PYPPBOXLauncher", "LAUNCH"))
         self.detector_label.setText(_translate("PYPPBOXLauncher", "Detector"))